cleaning up

+++ src/pokemon_menu.c
@@ -934,7 +934,7 @@ void debug_sub_80986AC(void)
 static void sub_808AE8C(void)
 {
     u8 i;
-    u16 arg = gSpecialVar_ItemId - (ITEM_TM01 % 256);
+    u8 arg = gSpecialVar_ItemId - (ITEM_TM01 % 256);
     for (i = 0; i < 6; i++)
     {
         if (GetMonData(&gPlayerParty[i], MON_DATA_SPECIES))
